\name{IWTomics-internal}
\title{IWTomics internal functions}
\description{IWTomics internal functions}
\keyword{internal}

\alias{nRegions,IWTomicsData-method}
\alias{nFeatures,IWTomicsData-method}
\alias{dim,IWTomicsData-method}
\alias{lengthRegions,IWTomicsData-method}
\alias{lengthFeatures,IWTomicsData-method}
\alias{resolution,IWTomicsData-method}
\alias{metadata,IWTomicsData-method}
\alias{regions,IWTomicsData-method}
\alias{features,IWTomicsData-method}
\alias{idRegions,IWTomicsData-method}
\alias{idFeatures,IWTomicsData-method}
\alias{nameRegions,IWTomicsData-method}
\alias{nameFeatures,IWTomicsData-method}
\alias{alignment,IWTomicsData-method}
\alias{testInput,IWTomicsData-method}
\alias{nTests,IWTomicsData-method}
\alias{idRegionsTest,IWTomicsData-method}
\alias{idRegionsTest,IWTomicsData,missing-method}
\alias{idRegionsTest,IWTomicsData,vector-method}
\alias{idFeaturesTest,IWTomicsData-method}
\alias{adjusted_pval,IWTomicsData-method}
\alias{adjusted_pval,IWTomicsData,missing,missing,missing-method}
\alias{adjusted_pval,IWTomicsData,vector,character,vector-method}
\alias{adjusted_pval,IWTomicsData,missing,character,vector-method}
\alias{adjusted_pval,IWTomicsData,vector,missing,vector-method}
\alias{adjusted_pval,IWTomicsData,missing,missing,vector-method}
\alias{adjusted_pval,IWTomicsData,vector,character,missing-method}
\alias{adjusted_pval,IWTomicsData,missing,character,missing-method}
\alias{adjusted_pval,IWTomicsData,vector,missing,missing-method}

\alias{plot.IWTomicsData}
\alias{plot,IWTomicsData,missing-method}
\alias{plot,IWTomicsData,ANY-method}

