\name{regionsFeatures_center}
\alias{regionsFeatures_center}
\docType{data}

\title{Example of \code{"IWTomicsData"} object with center alignment}

\description{
  Example of \code{"\link{IWTomicsData}"} object with center alignment, used to illustrate 
  usage of IWTomics package.
}

\usage{data(regionsFeatures_center)}

\value{
  \code{"IWTomicsData"} object.
}

\format{
  An object of class \code{"IWTomicsData"}, with four region datasets "Elements 1", "Elements2", 
  "Elements 3" and "Controls" with center alignment and two features "Feature 1" and "Feature 2". 
}

\examples{
data(regionsFeatures_center)
regionsFeatures_center
}
\keyword{datasets}
