\name{regionsFeatures_scale}
\alias{regionsFeatures_scale}
\docType{data}

\title{Example of \code{"IWTomicsData"} object with scale alignment}

\description{
  Example of \code{"\link{IWTomicsData}"} object with scale alignment, used to illustrate 
  usage of IWTomics package.
}

\usage{data(regionsFeatures_scale)}

\value{
  \code{"IWTomicsData"} object.
}

\format{
  An object of class \code{"IWTomicsData"}, with four region datasets "Elements 1", "Elements2", 
  "Elements 3" and "Controls" with scale alignment and two features "Feature 1" and "Feature 2". 
}

\examples{
data(regionsFeatures_scale)
regionsFeatures_scale
}
\keyword{datasets}
